wait_for_object_in_inbox does not need a return value: throw TimeoutException instead use float for time delays as in python time.sleep()

+++ src/feditest/protocols/__init__.py
@@ -650,5 +650,5 @@ class TimeoutException(RuntimeError):
     """
     A result has not arrived within the expected time period.
     """
-    def __init__(self, msg: str, timeout: int):
+    def __init__(self, msg: str, timeout: float):
         super().__init__(f'{ msg } (timeout: { timeout })')
